Increased font sizes and reworked heading proportions/styles.
I made it bigger because I disliked how small everything was. And additionally, I increased the prominence of the headings (by boldening them) and implemented some deeper headings.

Please note that these proportions (such as font sizes and margin widths) are very much provisional. More experimentation is to be done with proportioning.

Issue: Reducing the width of the browser will push the sidebar off the side a bit. Some possible solutions: 1) After about a browser width of 1000px, lock the left margin and allow the right margin to continue to collapse. 2) Instead of a sidebar, add menu buttons under the site title. 3) Just make the sidebar smaller or closer to the main content.